+++ features/de.bund.bfr.knime.fsklab.feature/CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Changed
+ - Change the packages file name from packageList.txt to packages.json!
 ## [1.3.0] - 2018-11-02
 
 ### Added